requirements.txt
line-bot-sdk==2.4.1
selenium==4.15.0
webdriver-manager==4.0.1
flask==2.3.3
gunicorn==21.2.0